+++ CHANGELOG.md
@@ -24,6 +24,8 @@ Notes
 ### New Features
 - **Add German (de), Spanish (`es`) localization by generative AI**
   - Since I have no knowledge of languages other than Japanese and English, translations may contain strange translations. If you notice any translations that need correction, I would appreciate it if you could send feedback on what to change and how!
+- **Add access to each tab in the “View” menu**
+  - Added menu items to the “View” menu in the menu bar to open the “Server,” “Model,” and “Chat” tabs. Press `⌘ (Command)` + `1` to open the “Server” tab, `⌘ (Command)` + `2` for the ‘Model’ tab, and `⌘ (Command)` + `3` for the “Chat” tab.
 
 ### Bug Fixes and Improvements
 - **Fix issue that `scrollEdgeEffect` did not work in chat view**